requirements.txt
streamlit
pandasd
joblib
statsmodels
numpy